config {config}	R Documentation

Manage Environment Specific Configuration Values

Description

Manage configuration values across multiple environments (e.g. development, test, production). Read values using a function that determines the current environment and returns the appropriate value.

Details

For additional details see the package website at https://github.com/rstudio/config.

[Package config version 0.3 Index]