is_active {config}	R Documentation

Test active configuration.

Description

Check whether a configuration is currently active.

Usage


is_active(config)

Arguments

config
Configuration name

Details

The name of the currently active configuration is read from the R_CONFIG_ACTIVE environment variable. If the variable is not defined then the "default" configuration is used.

To test for whether a configuration is active you should use the is_active() function rather than inspecting the environment variable directly (this is to so that tests remain valid if other means of specifying configurations are introduced in the future).

Value

Logical indicating whether the specified configuration is active

See Also

get()

[Package config version 0.3.2 Index]