Merge two configurations. Always use as config::merge().

Merge one configuration into another recursively.

Usage

merge(base_config, merge_config)

Arguments

base_config

Configuration to merge values into

merge_config

Configuration to merge values from

Value

Configuration which includes the values from merge_config merged into base_config.

Warning - Do not attach the package using library(config)

We strongly recommend you use config::get() rather than attaching the package using library(config).

In fact, we strongly recommend you never use library(config).

The underlying reason is that the get() and merge() functions in {config} will mask these functions with the same names in base R.

See also